Remove warnings with the `/` operator (#63)
* Update numpy.hpp

* Update numpy.hpp

* Update numpy.hpp

* Update numpy.hpp

* Update numpy.hpp

* Static cast the scalar to float

+++ numpy/include/numpy.hpp
@@ -211,7 +211,7 @@ public:
 
     template <typename U>
     auto binary_operator_truediv_impl(const U& other) const {
-        xt::xarray<float_> result = xt::cast<float_>(_array) / other;
+        xt::xarray<float_> result = xt::cast<float_>(_array) / static_cast<float_>(other);
         return ndarray<float_>(result);
     }
 
@@ -969,7 +969,7 @@ auto operator* (const U& scalar, const ndarray<T>& array) {
 template <typename T, typename U, typename = std::enable_if_t<!is_ndarray_v<U>>>
 auto operator/ (const U& scalar, const ndarray<T>& array) {
     xt::xarray<T> arr = array.get_array();
-    xt::xarray<float_> result = scalar / xt::cast<float_>(arr);
+    xt::xarray<float_> result = static_cast<float_>(scalar) / xt::cast<float_>(arr);
     return ndarray<float_>(result);
 }
